+++ CHANGELOG.md
@@ -1,7 +1,14 @@
-## [0.0.1] - 2020-07-23.
+## [0.0.3] - 2020-09-04.
 
-* A package to display the timeline with statuses and number.
+* Read.md changes.
+* Code refactored - created widget instead of function returing widget. 
 
 ## [0.0.2] - 2020-09-04.
 
 * Code refactor with naming convention.
+
+## [0.0.1] - 2020-07-23.
+
+* A package to display the timeline with statuses and number.
+
+